requirements.txt
streamlit
google-generativeai
PyPDF2
pandas
altair